requirements.txt
py-cord==2.0.0b5
python-dotenv==0.19.2
requests==2.26.0